Assignment 1 - tweet sentiment
This takes two arguments: sentiment file and a sample set of tweets
(JSON format) and returns a simple index for each one.  If a term is
not in the sentiment file, we assign a value of zero.  This code is
less fugly than what I'd write later (and in a hurry).